Moved page-level components into new directory
